Malicious URL detection using Artificial Neural Networks (ANNs)
The significance of automated systems for identifying and blocking malicious URLs cannot be overstated in the context of cybersecurity. The project involves several key components, starting with the collection of a labeled dataset containing examples of both malicious and benign URLs. Leveraging a dataset such as URL  can provide a solid foundation for model training. Training the ANN involves splitting the dataset into training and testing sets, optimizing hyperparameters, and ensuring the model generalizes well to unseen data. Evaluation metrics, including precision, recall, F1 score, and accuracy, guide the assessment of the model's performance. Once satisfied with the results, deployment becomes a focus, with considerations for creating a REST API for seamless integration into production environments.

Continuous monitoring is emphasized to keep the model updated in response to evolving threats. The project also provides clear instructions on dependencies, usage guidelines, and contribution pathways for community involvement. Ethical considerations and adherence to legal guidelines are paramount in the context of security-related projects. Additionally, the project encourages staying informed about cybersecurity developments to ensure the model remains adaptive to emerging threats.